+++ CHANGELOG.md
@@ -19,6 +19,34 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Removed
 
 
+## [2.5.0] - 2021-MM-DD
+
+### Security
+
+### Deprecated
+
+### Added
+
+- `Api.GeneralMerchantsApi.getPublicData()`
+- `Model.GeneralMerchantsPublicDataCompany`
+- `Model.GeneralMerchantsPublicDataModel`
+- `Model.SecupayTransactionCaptureDTO` element `shipping_information`
+
+### Changed
+
+- `Api.GeneralContractsApi.getAvailablePaymentMethods()` returns now a list of `string` instead of `StringList`
+
+### Fixed
+
+- `Model.GeoAddress` element `address_components` is a List of `AddressComponents`
+- `Model.PaymentCancelResult` element `result` is now an instance of `PaymentCancelResultDetails` instead of `object`
+- `Model.SecupayTransactionCancelDTO` element `contract` is now an instance of `PaymentContractsProductModel` instead of `object`
+- `Model.SecupayTransactionCaptureDTO` element `contract` is now an instance of `PaymentContractsProductModel` instead of `object`
+- `Model.SecupayTransactionUpdateBasketDTO` element `contract` is now an instance of `PaymentContractsProductModel` instead of `object`
+
+### Removed
+
+
 ## [2.4.0] - 2020-12-11
 
 ### Added
@@ -58,7 +86,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
-- `Model.SmartTransactionsDTO` element `device_source` is now a instance of `ProductInstanceUID` instead of `string`
+- `Model.SmartTransactionsDTO` element `device_source` is now an instance of `ProductInstanceUID` instead of `string`
 
 
 ## [2.2.0] - 2020-09-04
@@ -457,3 +485,4 @@ Beta version of the swagger generated PHP-SDK for the secuconnect API.
 [2.2.0]:https://github.com/secuconnect/secuconnect-php-sdk/compare/2.1.0...2.2.0
 [2.3.0]:https://github.com/secuconnect/secuconnect-php-sdk/compare/2.2.0...2.3.0
 [2.4.0]:https://github.com/secuconnect/secuconnect-php-sdk/compare/2.3.0...2.4.0
+[2.5.0]:https://github.com/secuconnect/secuconnect-php-sdk/compare/2.4.0...2.5.0